entity setup
