+++ restsecurity/src/main/java/dk/connectly/daos/IDAO.java
@@ -1,0 +1,20 @@
+package dk.connectly.daos;
+
+import dk.connectly.exceptions.ApiException;
+
+import java.util.List;
+
+public interface IDAO<T, K> {
+
+    List<T> getAll() throws ApiException;
+
+    T getById(K id) throws ApiException;
+
+    T create(T in) throws ApiException;
+
+    T update(T in, K id) throws ApiException;
+
+    T delete(K id) throws ApiException;
+
+
+}
