+++ restsecurity/src/main/java/dk/connectly/dtos/DTO.java
@@ -1,0 +1,8 @@
+package dk.connectly.dtos;
+
+public interface DTO<T> {
+
+    public T getId();
+    public void setId(T id);
+
+}
