+++ restsecurity/src/main/java/dk/connectly/dtos/TestDTO.java
@@ -1,0 +1,28 @@
+package dk.connectly.dtos;
+
+import lombok.*;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class TestDTO implements DTO<Integer> {
+    Integer id;
+    String name;
+
+    public TestDTO(String name) {
+        this.name = name;
+    }
+
+
+    @Override
+    public Integer getId() {
+        return id;
+    }
+
+    @Override
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+}
