+++ restsecurity/src/main/java/dk/connectly/dtos/TokenDTO.java
@@ -1,0 +1,17 @@
+package dk.connectly.dtos;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class TokenDTO {
+
+    private String token;
+    private String username;
+
+
+}
