+++ restsecurity/src/main/java/dk/connectly/exceptions/ApiException.java
@@ -1,0 +1,15 @@
+package dk.connectly.exceptions;
+
+import io.javalin.http.HttpStatus;
+import lombok.Getter;
+
+@Getter
+public class ApiException extends Exception{
+    private HttpStatus status;
+    public ApiException(int code, String msg){
+        super(msg+", "+code);
+        status = HttpStatus.forStatus(code);
+    }
+
+
+}
