+++ restsecurity/src/main/java/dk/connectly/exceptions/NotAuthorizedException.java
@@ -1,0 +1,18 @@
+package dk.connectly.exceptions;
+
+import lombok.Getter;
+
+@Getter
+public class NotAuthorizedException extends Exception {
+    private final int statusCode;
+
+    public NotAuthorizedException(int statusCode, String message) {
+        super(message);
+        this.statusCode = statusCode;
+    }
+    public NotAuthorizedException(int statusCode, String message, Throwable cause) {
+        super(message, cause);
+        this.statusCode = statusCode;
+    }
+
+}
