+++ restsecurity/src/main/java/dk/connectly/exceptions/ValidationException.java
@@ -1,0 +1,8 @@
+package dk.connectly.exceptions;
+
+public class ValidationException extends Throwable{
+
+    public ValidationException(String msg){
+        super(msg);
+    }
+}
