+++ restsecurity/src/main/java/dk/connectly/utils/ConnectionType.java
@@ -1,0 +1,9 @@
+package dk.connectly.utils;
+
+public enum ConnectionType {
+
+    FRIEND,
+    WORK,
+    FAMILY,
+    PUBLIC
+}
