+++ restsecurity/src/main/java/dk/connectly/utils/TokenUtils.java
@@ -1,0 +1,92 @@
+package dk.connectly.utils;
+
+import com.nimbusds.jose.*;
+import com.nimbusds.jose.crypto.MACSigner;
+import com.nimbusds.jose.crypto.MACVerifier;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
+import dk.connectly.dtos.UserDTO;
+import dk.connectly.exceptions.ApiException;
+import dk.connectly.exceptions.NotAuthorizedException;
+
+import java.text.ParseException;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+public class TokenUtils {
+
+    public UserDTO getUserWithRolesFromToken(String token) throws ParseException {
+        // Return a user with Set of roles as strings
+        SignedJWT jwt = SignedJWT.parse(token);
+        String roles = jwt.getJWTClaimsSet().getClaim("roles").toString();
+        String username = jwt.getJWTClaimsSet().getClaim("username").toString();
+
+        Set<String> rolesSet = Arrays
+                .stream(roles.split(","))
+                .collect(Collectors.toSet());
+        return new UserDTO(username, rolesSet);
+    }
+    public boolean tokenIsValid(String token, String secret) throws ParseException, JOSEException, NotAuthorizedException {
+        SignedJWT jwt = SignedJWT.parse(token);
+        if (jwt.verify(new MACVerifier(secret)))
+            return true;
+        else
+            throw new NotAuthorizedException(403, "Token is not valid");
+    }
+
+    public boolean tokenNotExpired(String token) throws ParseException, NotAuthorizedException {
+        if (timeToExpire(token) > 0)
+            return true;
+        else
+            throw new NotAuthorizedException(403, "Token has expired");
+    }
+    public int timeToExpire(String token) throws ParseException, NotAuthorizedException {
+        SignedJWT jwt = SignedJWT.parse(token);
+        return (int) (jwt.getJWTClaimsSet().getExpirationTime().getTime() - new Date().getTime());
+    }
+
+    public String createToken(UserDTO user, String ISSUER, String TOKEN_EXPIRE_TIME, String SECRET_KEY) throws ApiException {
+        // https://codecurated.com/blog/introduction-to-jwt-jws-jwe-jwa-jwk/
+        try {
+            JWTClaimsSet claimsSet = new JWTClaimsSet.Builder()
+                    .subject(user.getUsername())
+                    .issuer(ISSUER)
+                    .claim("username", user.getUsername())
+                    .claim("roles", user.getRoles().stream().reduce((s1, s2) -> s1 + "," + s2).get())
+                    .expirationTime(new Date(new Date().getTime() + Integer.parseInt(TOKEN_EXPIRE_TIME)))
+                    .build();
+            Payload payload = new Payload(claimsSet.toJSONObject());
+
+            JWSSigner signer = new MACSigner(SECRET_KEY);
+            JWSHeader jwsHeader = new JWSHeader(JWSAlgorithm.HS256);
+            JWSObject jwsObject = new JWSObject(jwsHeader, payload);
+            jwsObject.sign(signer);
+            return jwsObject.serialize();
+
+        } catch (JOSEException e) {
+            e.printStackTrace();
+            throw new ApiException(500, "Could not create token");
+        }
+    }
+
+
+    //CHECKS IF DEPLOYED
+    public String createToken(UserDTO user) throws ApiException{
+        String ISSUER;
+        String TOKEN_EXPIRE_TIME;
+        String SECRET_KEY;
+
+        if(System.getenv("DEPLOYED") != null){
+            ISSUER = System.getenv("ISSUER");
+            TOKEN_EXPIRE_TIME = System.getenv("TOKEN_EXPIRE_TIME");
+            SECRET_KEY = System.getenv("SECRET_KEY");
+        }else {
+            ISSUER = "Nicklas W.";
+            TOKEN_EXPIRE_TIME = "1800000";
+            SECRET_KEY = "ghjyhtgrfeghjyhtgrfeghjyhtgrfeghjyhtgrfeghjyhtgrfeghjyhtgrfeghjyhtgrfeghjyhtgr";
+        }
+        return createToken(user,ISSUER,TOKEN_EXPIRE_TIME, SECRET_KEY);
+    }
+}
