--- restsecurity/src/main/java/org/example/config/Routes.java
@@ -1,51 +1,0 @@
-package org.example.config;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.javalin.apibuilder.EndpointGroup;
-import io.javalin.security.RouteRole;
-import jakarta.persistence.EntityManagerFactory;
-import org.example.controllers.SecurityController;
-import org.example.controllers.TestController;
-import org.example.daos.TestMemoryDao;
-
-import static io.javalin.apibuilder.ApiBuilder.*;
-
-public class Routes {
-    private static SecurityController sc;
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-    private static TestController tc;
-
-    public static EndpointGroup getRoutes(boolean isTesting){
-        sc = SecurityController.getInstance(isTesting);
-        return () -> {
-            path("/", () -> {
-                get("/", ctx -> ctx.json(objectMapper.createObjectNode().put("Message", "Connected Successfully")), roles.ANYONE);
-            });
-            path("/auth", () -> {
-                post("/login", sc.login(), roles.ANYONE);
-                post("/register", sc.register(), roles.ANYONE);
-            });
-            path("/protected", () -> {
-                before(sc.authenticate());
-                get("/user_demo", ctx-> ctx.json(objectMapper.createObjectNode()), roles.USER);
-                get("/admin_demo", ctx-> ctx.json(objectMapper.createObjectNode()), roles.ADMIN);
-            });
-        };
-    }
-
-    public static EndpointGroup getTestRoutes(){
-        tc = TestController.getInstance();
-        return ()-> {
-            get("/getAll", tc.getAllTest());
-            post("/create",tc.createTest());
-        };
-    }
-
-
-    public enum roles implements RouteRole {
-        USER,
-        ADMIN,
-        ANYONE
-    }
-
-}
