--- restsecurity/src/main/java/org/example/daos/IDAO.java
@@ -1,20 +1,0 @@
-package org.example.daos;
-
-import org.example.exceptions.ApiException;
-
-import java.util.List;
-
-public interface IDAO<T, K> {
-
-    List<T> getAll() throws ApiException;
-
-    T getById(K id) throws ApiException;
-
-    T create(T in) throws ApiException;
-
-    T update(T in, K id) throws ApiException;
-
-    T delete(K id) throws ApiException;
-
-
-}
