--- restsecurity/src/main/java/org/example/daos/TestMemoryDao.java
@@ -1,17 +1,0 @@
-package org.example.daos;
-
-import org.example.dtos.TestDTO;
-
-public class TestMemoryDao extends MemoryDAO<TestDTO, Integer> {
-
-    private static TestMemoryDao instance;
-
-    public static TestMemoryDao getInstance() {
-        if (instance == null) {
-            instance = new TestMemoryDao();
-        }
-        return instance;
-    }
-
-
-}
