--- restsecurity/src/main/java/org/example/dtos/DTO.java
@@ -1,8 +1,0 @@
-package org.example.dtos;
-
-public interface DTO<T> {
-
-    public T getId();
-    public void setId(T id);
-
-}
