--- restsecurity/src/main/java/org/example/dtos/TestDTO.java
@@ -1,30 +1,0 @@
-package org.example.dtos;
-
-import lombok.*;
-
-import java.util.Objects;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class TestDTO implements DTO<Integer> {
-    Integer id;
-    String name;
-
-    public TestDTO(String name) {
-        this.name = name;
-    }
-
-
-    @Override
-    public Integer getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-}
