--- restsecurity/src/main/java/org/example/dtos/TokenDTO.java
@@ -1,17 +1,0 @@
-package org.example.dtos;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class TokenDTO {
-
-    private String token;
-    private String username;
-
-
-}
