--- restsecurity/src/main/java/org/example/exceptions/ValidationException.java
@@ -1,8 +1,0 @@
-package org.example.exceptions;
-
-public class ValidationException extends Throwable{
-
-    public ValidationException(String msg){
-        super(msg);
-    }
-}
